requirements.txt
gunicorn
gevent
flask
pdf2image
pillow
pypinyin